requirements.txt
mysql-connector-python==8.1.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
openpyxl==3.1.2